[dbsystel] Add pom.xml
Usually, Gradle’s `maven-publish` plug-in is responsible for generating
a POM file dynamically at build time.

However, thanks to the excellent work of @MrDOS and others, upstream's
Gradle-6-based process is currently being replaced [1] by a
Gradle-8-based one. The latter uses a different, more convenient method
to generate the POM.

In DB Systel’s fork, we’re not going to adopt that Gradle-8-based
build process for production until upstream PR #238 is properly tested,
merged, and included in a stable upstream release.

Given that we need to get two critical bugfixes that we’ve cherry-picked
into production quickly, we’re going to accept some duplication and
commit the POM to Git for the time being.

[1]: https://github.com/NeuronRobotics/nrjavaserial/pull/238