[dbsystel] Fix segfault if fd unavailable
The `get_java_var_long` function returns 0 in several failure modes,
e.g. if a file descriptor is unavailable. [1]

However, one of the call sites is missing the result check, which causes
a JVM segfault if the return value is 0. The segfault occurs on
dereferencing the pointer: [2]

```c
eis->eventflags[SPE_DATA_AVAILABLE]
```

Add a result value check, throwing a proper IOException if it is 0.

See also similar issue #59. [3]

Fixes #112 [4], #136 [5] and #242 [6].

[1]: https://github.com/NeuronRobotics/nrjavaserial/blob/0df8b60485a56d7698b71183237b5615d02a8194/src/main/c/src/SerialImp.c#L5137-L5142
[2]: https://github.com/NeuronRobotics/nrjavaserial/blob/0df8b60485a56d7698b71183237b5615d02a8194/src/main/c/src/SerialImp.c#L3085
[3]: https://github.com/NeuronRobotics/nrjavaserial/issues/59
[4]: https://github.com/NeuronRobotics/nrjavaserial/issues/112
[5]: https://github.com/NeuronRobotics/nrjavaserial/issues/136
[6]: https://github.com/NeuronRobotics/nrjavaserial/issues/242

Reported-by: Alex Vasiliev <@alex-vas>
Reported-by: Łukasz Dywicki <luke@code-house.org>
Reported-by: Jose Pacelli <josepacelli@gmail.com>
Reported-by: Frank Hartwig <frank.fr.hartwig@deutschebahn.com>