Unblock main thread if monitor thread init fails
When the monitor thread encounters an error during initialization,
the main thread hangs indefinitely because the monitor thread will bail
without ever releasing the synchronization lock and without signaling
the `Condition` on which the main thread is waiting.

Unblock the main thread by adding the signal/unlock sequence to the
failure path.

Additionally, throw a Java exception and propagate it to the main thread
so the caller gets to see exactly what caused the error.

Co-authored-by: Jan-Niklas Dornbach <jan-niklas.dornbach@deutschebahn.com>
Co-authored-by: Martin Mungard <martin.mungard-extern@deutschebahn.com>
Co-authored-by: Richard L. Jackson <richard.r.jackson-extern@deutschebahn.com>